+++ src/ERPAccounting.Domain/Lookups/LookupModels.cs
@@ -61,11 +61,11 @@ public record ArticleLookup(
     [property: Column("NAZIV ARTIKLA")] string NazivArtikla,
     [property: Column("JM")] string? JedinicaMere,
     [property: Column("IDPoreskaStopa")] string? IdPoreskaStopa,
-    [property: Column("ProcenatPoreza")] decimal ProcenatPoreza,
-    [property: Column("Akciza")] decimal Akciza,
-    [property: Column("KoeficijentKolicine")] decimal KoeficijentKolicine,
+    [property: Column("ProcenatPoreza")] double ProcenatPoreza,
+    [property: Column("Akciza")] double Akciza,
+    [property: Column("KoeficijentKolicine")] double KoeficijentKolicine,
     [property: Column("ImaLot")] bool ImaLot,
-    [property: Column("OtkupnaCena")] decimal? OtkupnaCena,
+    [property: Column("OtkupnaCena")] double? OtkupnaCena,
     [property: Column("PoljoprivredniProizvod")] bool PoljoprivredniProizvod
 );
